fixed rounding

+++ app/controllers/calculations_controller.rb
@@ -35,7 +35,7 @@ class CalculationsController < ApplicationController
      render("calculations/square_form_template.html.erb")    
       end
   def process_square
-      @user_number = params["the_user_number"].to_i
+      @user_number = params["the_user_number"].to_f
       @squared_number = @user_number**2
      render ("calculations/square_results_template.html.erb") 
   end
@@ -45,7 +45,7 @@ class CalculationsController < ApplicationController
   end
   
   def process_square_root
-      @user_number = params["the_user_number"].to_i
+      @user_number = params["the_user_number"].to_f
       @square_root = @user_number**0.5
       render ("calculations/square_root_results_template.html.erb")
   end
